refactor: voucher 할당 서비스 로직을 wallet controller/service에서 실행하도록 리팩터링
