+++ src/main/java/devcourse/springbootbasic/commandline/function/FunctionHandler.java
@@ -6,12 +6,15 @@ import devcourse.springbootbasic.commandline.constant.InputMessage;
 import devcourse.springbootbasic.commandline.domain.VoucherTypeSelector;
 import devcourse.springbootbasic.controller.CustomerController;
 import devcourse.springbootbasic.controller.VoucherController;
+import devcourse.springbootbasic.controller.WalletController;
 import devcourse.springbootbasic.domain.voucher.VoucherType;
 import devcourse.springbootbasic.dto.customer.CustomerCreateRequest;
 import devcourse.springbootbasic.dto.customer.CustomerFindResponse;
 import devcourse.springbootbasic.dto.customer.CustomerResponse;
 import devcourse.springbootbasic.dto.customer.CustomerUpdateBlacklistRequest;
 import devcourse.springbootbasic.dto.voucher.*;
+import devcourse.springbootbasic.dto.wallet.VoucherAssignRequest;
+import devcourse.springbootbasic.dto.wallet.VoucherAssignResponse;
 import devcourse.springbootbasic.exception.InputErrorMessage;
 import devcourse.springbootbasic.exception.InputException;
 import lombok.RequiredArgsConstructor;
@@ -28,6 +31,7 @@ public class FunctionHandler {
 
     private final VoucherController voucherController;
     private final CustomerController customerController;
+    private final WalletController walletController;
     private final ConsoleIOHandler consoleIOHandler;
 
     public void exit() {
@@ -95,7 +99,7 @@ public class FunctionHandler {
         UUID voucherId = consoleIOHandler.inputUUIDWithMessage(InputMessage.VOUCHER_ID);
         UUID customerId = consoleIOHandler.inputUUIDWithMessage(InputMessage.CUSTOMER_ID);
 
-        VoucherAssignResponse voucher = voucherController.assignVoucherToCustomer(new VoucherAssignRequest(voucherId, customerId));
+        VoucherAssignResponse voucher = walletController.assignVoucherToCustomer(new VoucherAssignRequest(voucherId, customerId));
         log.info(String.format(ConsoleConstants.VOUCHER_ASSIGN_MESSAGE, voucher.getVoucherId(), voucher.getCustomerId()));
     }
 }
